+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v0.15.9-SNAPSHOT
+* Terminate `ServerApp` even if the server fails to start
 * Patch-level upgrades to dependencies:
     * async-http-client-2.0.31
     * jetty-9.3.18.v20170406